+++ HISTORY.md
@@ -33,6 +33,7 @@
 
 #### Improvements
 
+- Use dtypes when calling asarray in archive methods ({pr}`673`)
 - Replace cKDTree usage with KDTree ({pr}`669`)
 - Support ProximityArchive in parallel_axes_plot ({pr}`647`)
 - Cast dtype when validating arguments ({pr}`646`)